requirements.txt
pymupdf
streamlit
google-genai
arabic-reshaper
python-bidi
